LayoutTests/imported/w3c/web-platform-tests/trusted-types/block-string-assignment-to-Range-createContextualFragment-expected.txt


PASS range.createContextualFragment assigned via policy (successful HTML transformation).
FAIL `range.createContextualFragment(string)` throws. assert_throws_js: function "_ => {
      var result = range.createContextualFragment("A string");
    }" did not throw
FAIL `range.createContextualFragment(null)` throws. assert_throws_js: function "_ => {
      var result = range.createContextualFragment(null);
    }" did not throw
FAIL `range.createContextualFragment(string)` assigned via default policy (successful HTML transformation). assert_equals: expected "Quack, I want to be a duck!" but got "Hi, I want to be transformed!"
PASS `range.createContextualFragment(null)` assigned via default policy does not throw.
